+++ hy/__init__.py
@@ -12,10 +12,10 @@ def _initialize_env_var(env_var, default_val):
     return bool(os.environ.get(env_var, default_val))
 
 
-import hy.importer  # NOQA
-
+# Import for side-effects.
+import hy.importer, hy.hy_inspect
 hy.importer._inject_builtins()
-# we import for side-effects.
+
 
 
 class I: